Another set of compile warnings (observed when compiling with MSVC)
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8557 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/lts/source/exploration.cpp
@@ -648,7 +648,7 @@ namespace mcrl2
       }
       else
       {
-        std::pair<long, bool> result = states.put(state);
+        std::pair<size_t, bool> result = states.put(state);
         is_new = result.second;
         return result.first;
       }
